frontend: Extract getUserIdFromLocalStorage to new file
