+++ frontend/src/lib/k8s/api/v1/streamingApi.ts
@@ -15,8 +15,8 @@
  */
 
 import { isDebugVerbose } from '../../../../helpers/debugVerbose';
-import { getUserIdFromLocalStorage } from '../../../../stateless';
 import { findKubeconfigByClusterName } from '../../../../stateless/findKubeconfigByClusterName';
+import { getUserIdFromLocalStorage } from '../../../../stateless/getUserIdFromLocalStorage';
 import { getCluster } from '../../../cluster';
 import { KubeObjectInterface } from '../../KubeObject';
 import { ApiError } from '../v2/ApiError';
