+++ frontend/src/lib/k8s/api/v2/fetch.test.ts
@@ -16,8 +16,8 @@
 
 import nock from 'nock';
 import { afterEach, beforeEach, describe, expect, it, Mock, vi } from 'vitest';
-import { getUserIdFromLocalStorage } from '../../../../stateless';
 import { findKubeconfigByClusterName } from '../../../../stateless/findKubeconfigByClusterName';
+import { getUserIdFromLocalStorage } from '../../../../stateless/getUserIdFromLocalStorage';
 import { getClusterAuthType } from '../v1/clusterRequests';
 import { BASE_HTTP_URL, clusterFetch } from './fetch';
 
@@ -30,7 +30,7 @@ vi.mock('../../../../stateless/findKubeconfigByClusterName', () => ({
   findKubeconfigByClusterName: vi.fn(),
 }));
 
-vi.mock('../../../../stateless', () => ({
+vi.mock('../../../../stateless/getUserIdFromLocalStorage', () => ({
   getUserIdFromLocalStorage: vi.fn(),
 }));
