+++ frontend/src/lib/k8s/api/v2/webSocket.test.ts
@@ -17,8 +17,8 @@
 import { renderHook } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import WS from 'vitest-websocket-mock';
-import { getUserIdFromLocalStorage } from '../../../../stateless';
 import { findKubeconfigByClusterName } from '../../../../stateless/findKubeconfigByClusterName';
+import { getUserIdFromLocalStorage } from '../../../../stateless/getUserIdFromLocalStorage';
 import { getCluster } from '../../../cluster';
 import { BASE_WS_URL, MULTIPLEXER_ENDPOINT, useWebSocket, WebSocketManager } from './webSocket';
 
@@ -31,7 +31,7 @@ vi.mock('../../../../stateless/findKubeconfigByClusterName', () => ({
   findKubeconfigByClusterName: vi.fn(),
 }));
 
-vi.mock('../../../../stateless', () => ({
+vi.mock('../../../../stateless/getUserIdFromLocalStorage', () => ({
   getUserIdFromLocalStorage: vi.fn(),
 }));
