+++ frontend/src/stateless/getUserIdFromLocalStorage.ts
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2025 The Kubernetes Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { generateSecureToken } from '.';
+
+/**
+ * In the backend we use a unique ID to identify a user. If there is no ID in localStorage
+ * we generate a new one and store it in localStorage. We then combine it with the
+ * cluster name and this headlamp-userId to create a unique ID for a cluster. If we don't
+ * do it then if 2 different users have a cluster with the same name, then the
+ * proxy will be overwritten.
+ * @returns headlamp-userId from localStorage
+ */
+
+export function getUserIdFromLocalStorage(): string {
+  let headlampUserId = localStorage.getItem('headlamp-userId');
+
+  if (!headlampUserId) {
+    headlampUserId = generateSecureToken();
+
+    if (headlampUserId) {
+      localStorage.setItem('headlamp-userId', headlampUserId);
+    }
+  }
+
+  return headlampUserId!;
+}
